+++ src/plugins/relay/relay-config.h
@@ -70,9 +70,11 @@ extern struct t_config_option *relay_config_irc_backlog_since_last_message;
 extern struct t_config_option *relay_config_irc_backlog_tags;
 extern struct t_config_option *relay_config_irc_backlog_time_format;
 
-extern struct t_config_option *relay_config_api_remote_get_lines;
 extern struct t_config_option *relay_config_api_remote_autoreconnect_delay_growing;
 extern struct t_config_option *relay_config_api_remote_autoreconnect_delay_max;
+extern struct t_config_option *relay_config_api_remote_get_lines;
+extern struct t_config_option *relay_config_api_remote_input_cmd_local;
+extern struct t_config_option *relay_config_api_remote_input_cmd_remote;
 
 extern int relay_config_auto_open_buffer[];
 extern int relay_config_display_clients[];